Add Null ptr checks to avoid crash when system table is empty
There was a crash in lldpd that was uncovered when running tests
to check regression with the changes for OVS 2.5 upgrade. From
the backtrace it appeared that we are running into a case where
a daemon starts processing before its IDL cache has got a chance
to get updated. These NULL ptr checks protect against this case.

Change-Id: I7f87e7ff3a95f6dfc48ead29e96c496e104a9b6b
Signed-off-by: Shad Ansari <shad.ansari@hpe.com>